fix continue button
