+++ js/src/components/paid-ads/ads-campaign/ads-campaign.js
@@ -102,7 +102,6 @@ export default function AdsCampaign( {
 					{ typeof continueButton === 'function'
 						? continueButton( {
 								formProps: formContext,
-								trackingContext,
 						  } )
 						: continueButton }
 				</StepContentActions>
